fix: time in edges expressed in minutes

+++ src/model/edge.py
@@ -6,7 +6,7 @@ class Edge(db.Model):
     start_node_id= db.Column(db.Integer,db.ForeignKey('node.id'),nullable=False)
     end_node_id = db.Column(db.Integer,db.ForeignKey('node.id'),nullable=False)
     distance = db.Column(db.Integer, nullable=False)
-    time = db.Column(db.Float, nullable=False)
+    time = db.Column(db.Integer, nullable=False)
 
     start_node = relationship("Node", foreign_keys=[start_node_id])
     end_node = relationship("Node", foreign_keys=[end_node_id])
